requirements.txt
docker-py
netifaces
PyYAML
pbr
